Split C# static interface prepass
